Replace ProbablilisticTensorDictModule with prototype
