+++ .gitignore
@@ -3,3 +3,4 @@ __pycache__
 *.egg-info
 tensordict/version.py
 .pytest_cache
+build
